LayoutTests/inspector/debugger/break-on-exception-throw-in-promise-expected.txt

CONSOLE MESSAGE: testThrowingInPromise
CONSOLE MESSAGE: in promise
CONSOLE MESSAGE: throwing TestError
CONSOLE MESSAGE: Unhandled Promise Rejection: TestError
CONSOLE MESSAGE: testThrowingInPromiseThen
CONSOLE MESSAGE: in promise
CONSOLE MESSAGE: in promise.then
CONSOLE MESSAGE: throwing TestError
CONSOLE MESSAGE: Unhandled Promise Rejection: TestError
CONSOLE MESSAGE: testThrowingInPromiseThenWithCatch
CONSOLE MESSAGE: in promise
CONSOLE MESSAGE: in promise.then
CONSOLE MESSAGE: throwing TestError
CONSOLE MESSAGE: in promise.catch
CONSOLE MESSAGE: DONE
CONSOLE MESSAGE: testThrowingInPromiseWithCatch
CONSOLE MESSAGE: in promise
CONSOLE MESSAGE: throwing TestError
CONSOLE MESSAGE: in promise.catch
CONSOLE MESSAGE: DONE
CONSOLE MESSAGE: testThrowingInPromiseWithRethrowInCatch
CONSOLE MESSAGE: in promise
CONSOLE MESSAGE: throwing TestError
CONSOLE MESSAGE: in promise.catch
CONSOLE MESSAGE: Unhandled Promise Rejection: TestError
Checking pause locations within Promises when pausing on all exceptions.


== Running test suite: BreakOnAnyException.Promise
-- Running test case: BreakOnAnyException.Promise.ExceptionInPromiseConstructor
PASS: Should pause for exception.
PAUSE AT doThrow:8:22

-- Running test case: BreakOnAnyException.Promise.ExceptionInPromiseThenWithoutCatch
PASS: Should pause for exception.
PAUSE AT doThrow:8:22

-- Running test case: BreakOnAnyException.Promise.ExceptionInPromiseThenWithCatch
PASS: Should pause for exception.
PAUSE AT doThrow:8:22

-- Running test case: BreakOnAnyException.Promise.ExceptionInPromiseCatch
PASS: Should pause for exception.
PAUSE AT doThrow:8:22

-- Running test case: BreakOnAnyException.Promise.ExceptionInPromiseThenAndRethrownInCatch
PASS: Should pause for exception.
PAUSE AT doThrow:8:22
PASS: Should pause for exception.
PAUSE AT promiseCatch:86:16
